refactor(structure): reorganizar proyecto en paquetes por dominio
- Se crearon subpaquetes 'profugo' y 'cazador' en model para mayor claridad y separación de responsabilidades
- Se movieron las clases existentes a 'model.profugo'
- Preparado paquete 'cazador' para futuras clases
- Se actualizaron rutas de tests para reflejar la nueva estructura
